Write the rough outlines of the actual DriveByWire code.
It will boot up, check valet mode (hard-coded to *true* for testing).
If not valet mode, then verify the fingerprint.

Then wait for a button press.

Keep track of which gear is enabled, and if the same button is pressed
again, just flash its LED three times. Else, send a signal to the
actuator control task to change position of the gear selector on the
gearbox.

Starts up with button (P)ark enabled for testing. Eventually, we'll
"ask" the actuator what gear it is in and go from there.