Improve the double-button code.
The double-click didn't work properly. It detected BOTH buttons,
leading to 'N' still being active, so when the 'D' button was clicked,
it toggled the Valet Mode *again*. AND again! Took a while for the
system to "catch up".

Solve this by waiting a second after we've updated the flash.
+ Blink the 'D' LED three times, started a few miliseconds before
  the "same button pressed" ('N') LED starts blinking.
  So BOTH LEDs will blink, with a bit of offset. This to indicate
  that Valet Mode have been toggled.